backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
alembic